requirements.txt
flask
pandas
pymysql
sshtunnel
SQLAlchemy
marshmallow
python-dotenv
psycopg2-binary
Flask-SQLAlchemy